chức năng đánh giá cửa hàng

+++ src/main/java/Api/AppDatDoAn/dto/CuaHangDto.java
@@ -18,7 +18,7 @@ public class CuaHangDto {
 
     private Long luotdanhgia;
 
-    private int chatluong;
+    private double chatluong;
 
     private Date giomocua;
